+++ priv/repo/migrations/20201004093810_create_categories.exs
@@ -1,0 +1,13 @@
+defmodule Misobo.Repo.Migrations.CreateCategories do
+  use Ecto.Migration
+
+  def change do
+    create table(:categories) do
+      add :name, :string
+      add :desc, :string
+
+      timestamps()
+    end
+
+  end
+end