+++ lib/misobo_web/router.ex
@@ -15,6 +15,11 @@ defmodule MisoboWeb.Router do
 
     post("/account/signup", AccountController, :create)
     post("/account/login", AccountController, :login)
+
+    # pipe auth
+
+    put("/account/:id", AccountController, :update)
+    get("/categories", CategoryController, :index)
   end
 
   # Enables LiveDashboard only for development