+++ assets/js/app.js
@@ -12,4 +12,3 @@ import "../css/app.scss"
 //     import {Socket} from "phoenix"
 //     import socket from "./socket"
 //
-