+++ lib/misobo/communication/sms_provider/tesla/text_local_tesla.ex
@@ -1,0 +1,8 @@
+defmodule Misobo.Communication.SMSProvider.TextLocalTesla do
+  @moduledoc """
+  The TextLocal communication interface.
+  """
+  use Tesla
+  plug Tesla.Middleware.BaseUrl, "https://api.textlocal.in"
+  plug Tesla.Middleware.FormUrlencoded
+end