+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Removed for now removed features.
 Fixed for any bug fixes.
 Security in case of vulnerabilities. -->
 
+## [unreleased]
+- Added category-subcategory
+
+
 ## [0.1.0]
 
 - Added user-signup