+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Security in case of vulnerabilities. -->
 
 ## [0.1.0]
 
-- Added user-signup 
+- Added user-signup
 - Added the user login and token generation
 - Added the github-ci
 - Added the docker and docker-compose file