+++ CHANGELOG.md
@@ -13,14 +13,29 @@ Removed for now removed features.
 Fixed for any bug fixes.
 Security in case of vulnerabilities. -->
 
-## [unreleased]
+## [0.2.0]
+
+# Added
+
 - Added category-subcategory.
 - Added the setup for seeds data.
+- Added the seeds for cateogry
+- Added User SMS integration with vendor TextLocal
+- Added User login streak implementation
+- Added User Authentication implementation via Phoenix token
 
+
+## [0.1.1]
+
+### Fixed
+
+- Fixes for the DB related updates taking from sysenv
 
 ## [0.1.0]
 
-- Added user-signup 
+### Added
+
+- Added user-signup
 - Added the user login and token generation
 - Added the github-ci
 - Added the docker and docker-compose file